Move text_units and summarized entities
