+++ graphrag/index/workflows/v1/subflows/create_summarized_entities.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""All the steps to transform final documents."""
+"""All the steps to summarize entities."""
 
 from typing import Any, cast
 
@@ -15,11 +15,10 @@ from datashaper import (
 from datashaper.table_store.types import VerbResult, create_verb_result
 
 from graphrag.index.cache import PipelineCache
-from graphrag.index.storage import PipelineStorage
-from graphrag.index.verbs.entities.summarize.description_summarize import (
-    summarize_descriptions_df,
+from graphrag.index.flows.create_summarized_entities import (
+    create_summarized_entities as create_summarized_entities_flow,
 )
-from graphrag.index.verbs.snapshot_rows import snapshot_rows_df
+from graphrag.index.storage import PipelineStorage
 
 
 @verb(
@@ -36,26 +35,17 @@ async def create_summarized_entities(
     graphml_snapshot_enabled: bool = False,
     **_kwargs: dict,
 ) -> VerbResult:
-    """All the steps to transform final documents."""
+    """All the steps to summarize entities."""
     source = cast(pd.DataFrame, input.get_input())
 
-    summarized = await summarize_descriptions_df(
+    output = await create_summarized_entities_flow(
         source,
         cache,
         callbacks,
-        column="entity_graph",
-        to="entity_graph",
-        strategy=strategy,
-        num_threads=num_threads,
+        storage,
+        strategy,
+        num_threads,
+        graphml_snapshot_enabled,
     )
 
-    if graphml_snapshot_enabled:
-        await snapshot_rows_df(
-            summarized,
-            column="entity_graph",
-            base_name="summarized_graph",
-            storage=storage,
-            formats=[{"format": "text", "extension": "graphml"}],
-        )
-
-    return create_verb_result(cast(Table, summarized))
+    return create_verb_result(cast(Table, output))
